third_party/blink/web_tests/wpt_internal/display-lock/rendersubtree/activation/anchor-links-ancestor.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Display Locking: anchor links via click</title>
<link rel="author" title="Rakina Zata Amni" href="mailto:rakina@chromium.org">
<link rel="help" href="https://github.com/WICG/display-locking">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/utils.js"></script>
<div id=spacer style="height: 3000px"></div>
<div id="outermost">
  Outermost
  <div id="outer">
    Outer
    <div id="inner">
      Inner
      <div id="innermost">
        Innermost
      </div>
    </div>
  </div>
</div>
<a id="innermostLink" href ="#innermost">Click</a>
<script>
'use strict';
function prepareTest() {
  return Promise.all([
    setInvisibleActivatable(outer),
    setInvisibleActivatable(inner)
  ]);
}

promise_test(() => {
  return new Promise((resolve, reject) => {
    prepareTest().then(() => {
      assert_equals(outermost.renderSubtree, "");
      assert_equals(outer.renderSubtree, INVISIBLE_ACTIVATABLE);
      assert_equals(inner.renderSubtree, INVISIBLE_ACTIVATABLE);
      assert_equals(innermost.renderSubtree, "");

      let innerPromise = new Promise((resolve, reject) => {
        inner.onbeforeactivate = (e) => {
          assert_equals(e.activatedElement, innermost);
          resolve();
        }
      });

      let outerPromise = new Promise((resolve, reject) => {
        outer.onbeforeactivate = (e) => {
          assert_equals(e.activatedElement, innermost);
          // Resolve if this is directly targeted to #outer,
          // instead of bubbling here.
          if (e.target == outer)
            resolve();
        }
      });

      let outermostPromise = new Promise((resolve, reject) => {
        outermost.onbeforeactivate = (e) => {
          assert_equals(e.activatedElement, innermost);
          assert_not_equals(e.target, outermost);
          // Resolve if this is targeted to #outer, which is
          // dispatched after the event targeted to #inner.
          if (e.target == outer)
            resolve();
        }
      });

      innermost.onbeforeactivate = reject;
      // Navigating to element in locked subtree should fire beforeactivate
      // on locked ancestors, but not itself.
      innermostLink.click();
      Promise.all([innerPromise, outerPromise, outermostPromise]).then(resolve);
    });
  });
}, "Activation through anchor link fires beforeactivate on locked ancestor");
</script>
</html>
